requirements.txt
Flask-SQLAlchemy
sqlalchemy_utils
mysqlclient
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-script
flask-admin
flask-login